Backend

See merge request s08-blockchain-contract-sub2/S08P22A209!270

+++ backend/src/main/java/com/donjo/backend/api/dto/member/WishListItem.java
@@ -17,9 +17,9 @@ public class WishListItem {
     // 위시리스트 이미지 경로
     private String imgPath;
     // 위시리스트 지금까지 달성 수입
-    private Long collectedAmount;
+    private Double collectedAmount;
     // 위시리스트 목표 금액
-    private Long totalAmount;
+    private Double totalAmount;
     // wishlist DTO에 담기
     public static WishListItemBuilder builder(WishlistSol wishlistSol) {
         return MakeWishListItemBuilder()
